# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ServiceHub API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  provider: 'provider',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone: string;
  role: UserRole;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  user: 'user',
  provider: 'provider',
} as const;

export interface UserInput {
  name: string;
  email: string;
  phone: string;
  password: string;
  role: UserInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface LoginResult {
  user: User;
  token: string;
}

export type ProviderServiceType = typeof ProviderServiceType[keyof typeof ProviderServiceType];


export const ProviderServiceType = {
  electrician: 'electrician',
  plumber: 'plumber',
  carpenter: 'carpenter',
  ac_repair: 'ac_repair',
  tutor: 'tutor',
  mechanic: 'mechanic',
  beautician: 'beautician',
  delivery: 'delivery',
  freelancer: 'freelancer',
} as const;

export interface Provider {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  photo?: string | null;
  serviceType: ProviderServiceType;
  description: string;
  skills?: string[];
  rating: number;
  totalReviews: number;
  available: boolean;
  city: string;
  /** @nullable */
  distanceKm?: number | null;
  pricePerHour: number;
  /** @nullable */
  earningsTotal?: number | null;
  verified: boolean;
  createdAt: string;
}

export interface AvailabilityUpdate {
  available: boolean;
}

export interface VerifyProviderInput {
  verified: boolean;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  userId: number;
  providerId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userPhone?: string | null;
  /** @nullable */
  providerName?: string | null;
  serviceType: string;
  description: string;
  date: string;
  timeSlot: string;
  status: BookingStatus;
  amount: number;
  createdAt: string;
}

export interface BookingInput {
  userId: number;
  providerId: number;
  serviceType: string;
  description: string;
  date: string;
  timeSlot: string;
  amount: number;
  userName: string;
  userPhone: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface Review {
  id: number;
  /** @nullable */
  bookingId?: number | null;
  providerId: number;
  userId: number;
  /** @nullable */
  userName?: string | null;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  feedback: string;
  createdAt: string;
}

export interface ReviewInput {
  bookingId?: number;
  providerId: number;
  userId: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
  feedback: string;
}

export interface DayCount {
  date: string;
  count: number;
}

export interface AdminStats {
  totalUsers: number;
  totalProviders: number;
  totalBookings: number;
  bookingsToday: number;
  totalRevenue: number;
  bookingsByDay: DayCount[];
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  booking_confirmed: 'booking_confirmed',
  booking_request: 'booking_request',
  review_received: 'review_received',
  booking_completed: 'booking_completed',
} as const;

export interface Notification {
  id: number;
  userId: number;
  message: string;
  type: NotificationType;
  read: boolean;
  createdAt: string;
}

export interface EarningEntry {
  date: string;
  service: string;
  userName: string;
  amount: number;
  status: string;
}

export interface EarningsSummary {
  today: number;
  thisWeek: number;
  thisMonth: number;
  total: number;
  breakdown: EarningEntry[];
}

export interface Message {
  id: number;
  senderId: number;
  receiverId: number;
  /** @nullable */
  senderName?: string | null;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  senderId: number;
  receiverId: number;
  content: string;
}

export type ListProvidersParams = {
category?: string;
minRating?: number;
maxPrice?: number;
available?: boolean;
sortBy?: ListProvidersSortBy;
};

export type ListProvidersSortBy = typeof ListProvidersSortBy[keyof typeof ListProvidersSortBy];


export const ListProvidersSortBy = {
  rating: 'rating',
  distance: 'distance',
  price: 'price',
} as const;

export type ListBookingsParams = {
userId?: number;
providerId?: number;
status?: ListBookingsStatus;
};

export type ListBookingsStatus = typeof ListBookingsStatus[keyof typeof ListBookingsStatus];


export const ListBookingsStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export type ListReviewsParams = {
providerId?: number;
userId?: number;
};

export type ListNotificationsParams = {
userId?: number;
};

export type GetEarningsParams = {
providerId: number;
};

export type ListMessagesParams = {
userId?: number;
providerId?: number;
};