# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ServiceHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List service providers
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: available
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [rating, distance, price]
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get provider by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Provider not found
  /providers/{id}/availability:
    patch:
      operationId: updateProviderAvailability
      tags: [providers]
      summary: Toggle provider availability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvailabilityUpdate"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: providerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled]
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: providerId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /admin/providers/{id}/verify:
    patch:
      operationId: verifyProvider
      tags: [admin]
      summary: Verify/approve a service provider
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyProviderInput"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
...
components:
  schemas:
...
    Provider:
      type: object
      required: [id, userId, name, serviceType, description, rating, totalReviews, available, pricePerHour, city, verified, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        photo:
          type: ["string", "null"]
        serviceType:
          type: string
          enum: [electrician, plumber, carpenter, ac_repair, tutor, mechanic, beautician, delivery, freelancer]
        description:
          type: string
        skills:
          type: array
          items:
            type: string
        rating:
          type: number
        totalReviews:
          type: integer
        available:
          type: boolean
        city:
          type: string
        distanceKm:
          type: ["number", "null"]
        pricePerHour:
          type: number
        earningsTotal:
          type: ["number", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AvailabilityUpdate:
      type: object
      required: [available]
      properties:
        available:
          type: boolean
    VerifyProviderInput:
      type: object
      required: [verified]
      properties:
        verified:
          type: boolean
    Booking:
      type: object
      required: [id, userId, providerId, serviceType, description, date, timeSlot, status, amount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        providerId:
          type: integer
        userName:
          type: ["string", "null"]
        userPhone:
          type: ["string", "null"]
        providerName:
          type: ["string", "null"]
        serviceType:
          type: string
        description:
          type: string
        date:
          type: string
        timeSlot:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        amount:
          type: number
        createdAt:
          type: string
          format: date-time
    BookingInput:
      type: object
      required: [userId, providerId, serviceType, description, date, timeSlot, amount, userName, userPhone]
      properties:
        userId:
          type: integer
        providerId:
          type: integer
        serviceType:
          type: string
        description:
          type: string
        date:
          type: string
        timeSlot:
          type: string
        amount:
          type: number
        userName:
          type: string
        userPhone:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
    Review:
      type: object
      required: [id, providerId, userId, stars, feedback, createdAt]
      properties:
        id:
          type: integer
        bookingId:
          type: ["integer", "null"]
        providerId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        stars:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: string
        createdAt:
          type: string
          format: date-time
    ReviewInput:
      type: object
      required: [providerId, userId, stars, feedback]
      properties:
        bookingId:
          type: integer
        providerId:
          type: integer
        userId:
          type: integer
        stars:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: string
...