# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ServiceHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  userId: number;
  providerId: number;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userPhone?: string | null;
  /** @nullable */
  providerName?: string | null;
  serviceType: string;
  description: string;
  date: string;
  timeSlot: string;
  status: BookingStatus;
  amount: number;
  createdAt: Date;
}