# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ServiceHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  userId: number;
  providerId: number;
  serviceType: string;
  description: string;
  date: string;
  timeSlot: string;
  amount: number;
  userName: string;
  userPhone: string;
}