# lib/api-zod/src/generated/types/verifyProviderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ServiceHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyProviderInput {
  verified: boolean;
}